fix: remove helloworld base extension naming
