+++ src/integrationTest/java/com/hivemq/extensions/heartbeat/CustomConfigIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.helloworld;
+package com.hivemq.extensions.heartbeat;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import okhttp3.OkHttpClient;
@@ -29,7 +29,7 @@ import org.testcontainers.utility.MountableFile;
 
 import java.util.concurrent.TimeUnit;
 
-import static com.hivemq.extensions.helloworld.DockerImageNames.HIVEMQ;
+import static com.hivemq.extensions.heartbeat.DockerImageNames.HIVEMQ;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -38,25 +38,28 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
  * @since 1.0.4
  */
 @Testcontainers
-class DefaultConfigIT {
+class CustomConfigIT {
 
     @Container
     final @NotNull HiveMQContainer hivemq = new HiveMQContainer(HIVEMQ) //
             .withExtension(MountableFile.forClasspathResource("hivemq-heartbeat-extension"))
             .waitForExtension("HiveMQ Heartbeat Extension")
-            .withExposedPorts(9090)
+            .withExposedPorts(9191)
+            .withFileInExtensionHomeFolder(MountableFile.forClasspathResource("extension-config.xml"),
+                    "hivemq-heartbeat-extension",
+                    "/extension-config.xml")
             .withLogConsumer(outputFrame -> System.out.print("HiveMQ: " + outputFrame.getUtf8String()));
 
     @Test
     @Timeout(value = 2, unit = TimeUnit.MINUTES)
-    void defaultConfigFilePresent_defaultConfigFileUsed() throws Exception {
+    void customConfigPresent_customConfigUsed() throws Exception {
         final OkHttpClient client = new OkHttpClient();
 
         final Request request = new Request.Builder().url("http://" +
                 hivemq.getHost() +
                 ":" +
-                hivemq.getMappedPort(9090) +
-                "/heartbeat").build();
+                hivemq.getMappedPort(9191) +
+                "/custom-endpoint").build();
 
         try (final Response response = client.newCall(request).execute()) {
             final ResponseBody body = response.body();
