+++ src/integrationTest/java/com/hivemq/extensions/heartbeat/DockerImageNames.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.helloworld;
+package com.hivemq.extensions.heartbeat;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import org.testcontainers.utility.DockerImageName;
